+++ src/index.ts
@@ -450,11 +450,12 @@ export class WindowManager extends InvisiblePlugin<WindowMangerAttributes, any> 
                 for (const appId of apps) {
                     const appScenePath = appManager.store.getAppScenePath(appId);
                     if (appScenePath && appScenePath === scenePath) {
-                        console.warn(`[WindowManager]: ScenePath ${scenePath} Already opened`);
+                        console.warn(`[WindowManager]: ScenePath "${scenePath}" already opened`);
                         if (this.boxManager) {
                             const topBox = this.boxManager.getTopBox();
                             if (topBox) {
                                 this.boxManager.setZIndex(appId, topBox.zIndex + 1, false);
+                                this.boxManager.focusBox({ appId }, false);
                             }
                         }
                         return;