+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.4.66 (2023-12-12)
 
 - 修复: 窗口最大化时加载已存在 scenePath 的 app 没有自动切换 focus 的问题
 - 新增: `context.dispatchAppEvent(type, value?)` 发送 app 本地事件